requirements.txt
streamlit
pandas
numpy
opencv-python-headless
scikit-learn
scikit-image
Pillow
